+++ catcher/server/ws.go
@@ -1,0 +1,41 @@
+package server
+
+import (
+	"github.com/fasthttp/websocket"
+	"github.com/valyala/fasthttp"
+	"log"
+)
+
+var upgrader = websocket.FastHTTPUpgrader{
+	ReadBufferSize:  1024,
+	WriteBufferSize: 1024,
+	CheckOrigin: func(r *fasthttp.RequestCtx) bool {
+		return true
+	},
+}
+
+func catcherWebsocketsHandler(ctx *fasthttp.RequestCtx)  {
+	err := upgrader.Upgrade(ctx, func(conn *websocket.Conn) {
+		for {
+			mt, message, err := conn.ReadMessage()
+			if err != nil {
+				log.Println("read message error:", err)
+				break
+			}
+
+			log.Printf("recv: %s", message)
+
+			answerBuffer := []byte(processMessage(message).Message)
+			err = conn.WriteMessage(mt, answerBuffer)
+
+			if err != nil {
+				log.Println("write message error:", err)
+				break
+			}
+		}
+	})
+
+	if err != nil {
+		log.Println(err)
+	}
+}
